Split service into models, mappers, and service
Separate files for Nutriments, NutrientLevels, Product models.
Dedicated product-mapper with mapNutriments, mapNutrientLevels,
mapProduct functions. Service only handles fetch and delegates
mapping.
