+++ models/nutriments.ts
@@ -1,0 +1,11 @@
+export interface Nutriments {
+  energyKcal100g: number | null;
+  proteins100g: number | null;
+  carbohydrates100g: number | null;
+  fat100g: number | null;
+  saturatedFat100g: number | null;
+  sugars100g: number | null;
+  fiber100g: number | null;
+  salt100g: number | null;
+  sodium100g: number | null;
+}
